+++ CHANGELOG.md
@@ -1,8 +1,28 @@
+## 0.3
+
+Breaking changes:
+
+  - Materialized views are now manipulated through `materialize_view`,
+  `drop_view`, and `view_ready`
+
+  - Optional string parameter `enduser`, when creating an Empire
+  instance. This parameter is mandatory for performing materialized
+  view operations.
+
+Dependencies:
+
+  - New dependency on `[python-dateutil](https://labix.org/python-dateutil)`
+
+Features:
+
+  - `empire.view_materialized_at("viewname")` returns a `datetime`
+  object with the time the view was last materialized.
+
 ## 0.2
 
 Features:
 
-  - `query` returns Python objects
+  - `query` returns Python dicts
 
 License: